branch update
git-svn-id: https://svn.r-project.org/R/trunk@7081 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/scan.c
@@ -204,7 +204,7 @@ static int fillBuffer(char *buffer, SEXPTYPE type, int strip)
 donefill:
     if (strip)
     {
-        while (isspace(*--bufp))
+        while (isspace((int)*--bufp))
             ;
         bufp++;
     }
@@ -924,7 +924,7 @@ SEXP do_readln(SEXP call, SEXP op, SEXP args, SEXP rho)
         }
     }
     /* now strip white space off the end as well */
-    while (isspace(*--bufp))
+    while (isspace((int)*--bufp))
         ;
     *++bufp = '\0';
     ConsolePrompt[0] = '\0';
@@ -959,10 +959,10 @@ SEXP do_menu(SEXP call, SEXP op, SEXP args, SEXP rho)
     ConsolePrompt[0] = '\0';
 
     bufp = buffer;
-    while (isspace(*bufp))
+    while (isspace((int)*bufp))
         bufp++;
     first = LENGTH(CAR(args)) + 1;
-    if (isdigit(*bufp))
+    if (isdigit((int)*bufp))
     {
         first = strtod(buffer, NULL);
     }
